requirements.txt
python-a2s
trio==0.18.0
trio-typing==0.5.0
trio-paho-mqtt==0.3.1
